add nomic-cli and relayer subcommand
