+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "cli",
     "relayer",
     "core/primitives",
     "chain/client",
