+++ cli/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "nomic-cli"
+version = "0.1.0"
+authors = ["Judd <kepzorz@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { git = "https://github.com/clap-rs/clap/" }
+relayer = { path = "../relayer" }
